Make small tweaks
Remove the cron package, the garbage collection go routine is now called within main.

Remove unused struct from the database package.

Remove the dummy error handler.

Remove DefaultExpiryTime from the page struct.
