+++ .golangci.yaml
@@ -38,7 +38,6 @@ linters-settings:
           - github.com/redds-be/reddlinks/internal/http
           - github.com/redds-be/reddlinks/internal/json
           - github.com/redds-be/reddlinks/internal/utils
-          - github.com/redds-be/reddlinks/internal/cron
           - github.com/redds-be/reddlinks/test/helper
           - github.com/lib/pq
           - github.com/mattn/go-sqlite3
