--- internal/http/handler_err.go
@@ -1,29 +1,0 @@
-//    reddlinks, a simple link shortener written in Go.
-//    Copyright (C) 2024 redd
-//
-//    This program is free software: you can redistribute it and/or modify
-//    it under the terms of the GNU General Public License as published by
-//    the Free Software Foundation, either version 3 of the License, or
-//    (at your option) any later version.
-//
-//    This program is distributed in the hope that it will be useful,
-//    but WITHOUT ANY WARRANTY; without even the implied warranty of
-//    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//    GNU General Public License for more details.
-//
-//    You should have received a copy of the GNU General Public License
-//    along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-package http
-
-import (
-	"net/http"
-
-	"github.com/redds-be/reddlinks/internal/json"
-)
-
-// HandlerErr sends a dummy JSON error to the client.
-func HandlerErr(writer http.ResponseWriter, _ *http.Request) {
-	// Respond with a generic error at '/error'
-	json.RespondWithError(writer, http.StatusBadRequest, "Something went wrong.")
-}
