+++ test/http/test.tmpl
@@ -11,5 +11,5 @@
 <p>DefaultShortLength: {{.DefaultShortLength}}</p>
 <p>DefaultMaxShortLength: {{.DefaultMaxShortLength}}</p>
 <p>DefaultMaxCustomLength: {{.DefaultMaxCustomLength}}</p>
-<p>DefaultExpiryTime: {{.DefaultExpiryTime}}</p>
+<p>DefaultExpiryDate: {{.DefaultExpiryDate}}</p>
 <p>ContactEmail: {{.ContactEmail}}</p>
